requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
python-dotenv==0.9.1
requests==2.26.0
urllib3==1.26.7
